refactor(cli): listing inside an app is spelled apps files ls
`nib apps ls` sat one letter from `nib apps list` while meaning something
else entirely — one lists your apps, the other lists inside one. Moving
it under `files` says which is which by name, and gives the filesystem
commands somewhere to be added.

Both spellings move together: `apps files ls` and `apps files ls <path>`.
No behaviour changes.
