+++ packages/cli/src/command-tree.gen.ts
@@ -4,10 +4,10 @@ import type { InferForwardedOptions, InferParams, RuntimeNode } from '@parshjs/c
 import type { command as appsCmd } from './commands/apps.ts';
 import type { command as appsDeleteCmd } from './commands/apps/delete.ts';
 import type { command as appsExportDestinationCmd } from './commands/apps/export/[destination].ts';
+import type { command as appsFilesLsCmd } from './commands/apps/files/ls.ts';
+import type { command as appsFilesLsPathCmd } from './commands/apps/files/ls/[path].ts';
 import type { command as appsListCmd } from './commands/apps/list.ts';
 import type { command as appsLogsCmd } from './commands/apps/logs.ts';
-import type { command as appsLsCmd } from './commands/apps/ls.ts';
-import type { command as appsLsPathCmd } from './commands/apps/ls/[path].ts';
 import type { command as loginCmd } from './commands/login.ts';
 import type { command as runCommandCmd } from './commands/run/[command].ts';
 
@@ -29,6 +29,19 @@ declare module '@parshjs/core' {
       };
       rootOptions: {};
     };
+    'apps files ls': {
+      parents: {
+        'apps': { options: InferForwardedOptions<typeof appsCmd.options>; params: InferParams<typeof appsCmd.params> };
+      };
+      rootOptions: {};
+    };
+    'apps files ls [path]': {
+      parents: {
+        'apps': { options: InferForwardedOptions<typeof appsCmd.options>; params: InferParams<typeof appsCmd.params> };
+        'apps files ls': { options: InferForwardedOptions<typeof appsFilesLsCmd.options>; params: InferParams<typeof appsFilesLsCmd.params> };
+      };
+      rootOptions: {};
+    };
     'apps list': {
       parents: {
         'apps': { options: InferForwardedOptions<typeof appsCmd.options>; params: InferParams<typeof appsCmd.params> };
@@ -38,19 +51,6 @@ declare module '@parshjs/core' {
     'apps logs': {
       parents: {
         'apps': { options: InferForwardedOptions<typeof appsCmd.options>; params: InferParams<typeof appsCmd.params> };
-      };
-      rootOptions: {};
-    };
-    'apps ls': {
-      parents: {
-        'apps': { options: InferForwardedOptions<typeof appsCmd.options>; params: InferParams<typeof appsCmd.params> };
-      };
-      rootOptions: {};
-    };
-    'apps ls [path]': {
-      parents: {
-        'apps': { options: InferForwardedOptions<typeof appsCmd.options>; params: InferParams<typeof appsCmd.params> };
-        'apps ls': { options: InferForwardedOptions<typeof appsLsCmd.options>; params: InferParams<typeof appsLsCmd.params> };
       };
       rootOptions: {};
     };
@@ -90,6 +90,24 @@ export const commandTree: RuntimeNode = {
             paramChild: null,
           },
         },
+        'files': {
+          segment: { kind: 'literal', value: 'files' },
+          command: null,
+          literalChildren: {
+            'ls': {
+              segment: { kind: 'literal', value: 'ls' },
+              command: { path: 'apps files ls', load: () => import('./commands/apps/files/ls.ts').then((m) => m.command) },
+              literalChildren: {},
+              paramChild: {
+                segment: { kind: 'param', name: 'path' },
+                command: { path: 'apps files ls [path]', load: () => import('./commands/apps/files/ls/[path].ts').then((m) => m.command) },
+                literalChildren: {},
+                paramChild: null,
+              },
+            },
+          },
+          paramChild: null,
+        },
         'list': {
           segment: { kind: 'literal', value: 'list' },
           command: { path: 'apps list', load: () => import('./commands/apps/list.ts').then((m) => m.command) },
@@ -101,17 +119,6 @@ export const commandTree: RuntimeNode = {
           command: { path: 'apps logs', load: () => import('./commands/apps/logs.ts').then((m) => m.command) },
           literalChildren: {},
           paramChild: null,
-        },
-        'ls': {
-          segment: { kind: 'literal', value: 'ls' },
-          command: { path: 'apps ls', load: () => import('./commands/apps/ls.ts').then((m) => m.command) },
-          literalChildren: {},
-          paramChild: {
-            segment: { kind: 'param', name: 'path' },
-            command: { path: 'apps ls [path]', load: () => import('./commands/apps/ls/[path].ts').then((m) => m.command) },
-            literalChildren: {},
-            paramChild: null,
-          },
         },
       },
       paramChild: null,
