+++ packages/cli/src/commands/apps/files/ls.ts
@@ -7,12 +7,12 @@ import { listDirectory } from '#lib/filesystem.ts';
 import { isInteractive } from '#lib/ui.ts';
 
 /**
- * A command and the parent of `apps ls [path]` at once, which is how an optional positional is
- * spelled: the walk stops here when nothing follows, and the volume's root is what listing
+ * A command and the parent of `apps files ls [path]` at once, which is how an optional positional
+ * is spelled: the walk stops here when nothing follows, and the volume's root is what listing
  * without naming a directory means. `--deployment-id` is declared here and forwarded, so both
  * spellings take it the same way.
  */
-export const command = defineCommand('apps ls', {
+export const command = defineCommand('apps files ls', {
   description: "List a directory of an app's filesystem. Without a path, the volume root.",
   options: {
     'deployment-id': {
