+++ packages/cli/src/commands/apps/files/ls/[path].ts
@@ -5,7 +5,7 @@ import { requireSignedIn } from '#lib/credentials.ts';
 import { guestPath, listDirectory } from '#lib/filesystem.ts';
 import { isInteractive } from '#lib/ui.ts';
 
-export const command = defineCommand('apps ls [path]', {
+export const command = defineCommand('apps files ls [path]', {
   description: 'Directory to list, as a path inside the app filesystem.',
   options: {},
   params: {
@@ -29,7 +29,7 @@ export const command = defineCommand('apps ls [path]', {
     await listDirectory({
       api,
       slug,
-      deploymentId: parents['apps ls'].options['deployment-id'],
+      deploymentId: parents['apps files ls'].options['deployment-id'],
       path,
       print,
     });
